SIMSBIOHUB-30: Rename test file
